server/requirements.txt
asgiref==3.8.1
astroid==3.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.5.1
dill==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.1
psycopg2-binary==2.9.9
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.5.0
tomlkit==0.12.4
tzdata==2024.1
urllib3==2.2.1
